Refactor for Sysml Designer
- use uml designer core instead of uml designer plugin
