+++ plugins/org.obeonetwork.dsl.sysml.design/build.properties
@@ -15,4 +15,6 @@ bin.includes = META-INF/,\
                description/,\
                plugin.xml,\
                icons/,\
-               plugin.properties
+               plugin.properties,\
+               images/
+bin.excludes = description/sysmlOtherDiagram.odesign
