+++ tests/org.obeonetwork.dsl.sysml.design.tests/META-INF/MANIFEST.MF
@@ -25,7 +25,6 @@ Require-Bundle: org.eclipse.core.runtime,
  com.google.guava,
  org.eclipse.sirius.ui,
  org.eclipse.uml2.uml,
- org.obeonetwork.dsl.uml2.design,
  org.eclipse.sirius.diagram.sequence,
  org.eclipse.sirius.tests.swtbot.support
 Bundle-RequiredExecutionEnvironment: J2SE-1.5
